backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
boto3==1.36.2
botocore==1.36.2
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pydantic==2.10.5
pydantic_core==2.27.2
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.1
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
